extracting artists
